requirements.txt
flask-migrate
flask-restful
mysqlclient
requests
uwsgi
flask-redis
Flask-OpenID
